+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-06-08
+
 - Change signature of `parse`, `try_parse` to take `&PathBuf` instead of `PathBuf`
 - Add strict clippy settings
+- Add `watch` fn
 
 ## [0.3.2] - 2025-06-08
 
@@ -47,7 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The first release! Heavily WIP and has rough edges, but usable.
 
-[unreleased]: https://github.com/bahlo/pichu/compare/v0.3.2...HEAD
+[unreleased]: https://github.com/bahlo/pichu/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/bahlo/pichu/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/bahlo/pichu/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/bahlo/pichu/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/bahlo/pichu/compare/v0.2.0...v0.3.0